BAN-5 refactor: new structure Input
