+++ src/shared/ui/index.ts
@@ -7,7 +7,8 @@ export {Scroll} from './scroll/Scroll'
 export {Card} from './card/Card'
 export {Typography} from './typography/Typography'
 export {Recaptcha} from './recaptcha/Recaptcha'
-export {InputSlot,Input,type InputProps} from './input/Input'
+export {BaseInputSlot,BaseInput,type BaseInputProps} from './input/baseInput/BaseInput'
+export {Input,} from './input/Input'
 
 export {Navigation,NavMenuItem,NavMenuList} from './menu/Navigation'
 export {Menu} from "./menu/Menu";
