+++ src/shared/ui/input/Input.tsx
@@ -1,77 +1,41 @@
 'use client'
-import React, {type ComponentPropsWithoutRef, useId} from 'react';
-import clsx from "clsx";
-import s from './Input.module.scss'
-import {TextField} from "@radix-ui/themes";
-
-
-type InputTypes = 'text' | 'email' | 'password'
-
-export type InputProps = {
-    type?: InputTypes
-    subTitle?: string
-    error?: boolean
-    helperText?: string
-    side?: 'left' | 'right'
-
-
-} & Omit<ComponentPropsWithoutRef<typeof TextField.Root>, 'type'>
-
-type InputSlotProps = Omit<ComponentPropsWithoutRef<'button'>,'type'>
-
-export const Input = (props: InputProps) => {
+import React, {useState} from 'react';
+import {
+    BaseInput,
+    type BaseInputProps,
+    BaseInputSlot,
+    type BaseInputTypes,
+} from "@shared/ui/input/baseInput/BaseInput";
+import {EyeOffOutline, EyeOutline, SearchOutline} from "@/assets/icons/components";
 
 
 
-    const {
-        type = 'text',
-        subTitle = '',
-        error = false,
-        helperText = '',
-        side = 'right',
-        disabled = false,
-        className,
-        id,
-        children,
-        ["aria-label"]:ariaLabel,
-        ...rest
-    } = props
-
-
-    const idCurrent = id || useId()
-    const helperTextId = `${idCurrent}-helper-text`;
-    const labelId =`${idCurrent}-label`
-    const stateClassClsx = {
-        [s.disabled]: disabled,
-        [s.error]: error
-    };
-    const sideClassClsx = {
-        [s.left]: side === 'left',
-        [s.right]: side === 'right',
-    }
-    return <>
-        <TextField.Root className={clsx(s.wrapper,className, {...stateClassClsx},children && {...sideClassClsx})} type={type} id={idCurrent} disabled={disabled}
-                        aria-label={ariaLabel}
-                        aria-labelledby={subTitle && labelId}
-                        aria-describedby={helperText && helperTextId}
-                        aria-invalid={error}
-                        aria-disabled={disabled}
-                        {...rest} >
-
-            {children && <TextField.Slot className={clsx(s.children_wrapper, sideClassClsx)}>
-                {children}
-
-            </TextField.Slot>}
-            {subTitle && <label id={labelId} className={s.sub_title} htmlFor={idCurrent}>{subTitle}</label>}
-            {helperText && <span id={helperTextId} className={s.error_message}>{helperText}</span>}
-        </TextField.Root>
-
-    </>
+type Props = Omit<BaseInputProps, 'type'> & {
+    type?:BaseInputTypes |'search',
 }
 
-export const InputSlot =(props:InputSlotProps)=>{
-const {className,...rest} = props
+export const Input = (props: Props) => {
+    const [isShowPassword, setIsShowPassword] = useState(false)
 
-    return <button  className={clsx(s.slot_button,className)} type={'button'} {...rest} />
+    const {type='text',...rest} = props
 
-};
+
+
+    return <>
+        {type === 'text'||type === 'email' && <BaseInput type={type} {...rest} />}
+
+        {type === 'password' && <BaseInput type={isShowPassword ? 'text' : 'password'} {...rest} subTitle={'Password confirmation'}  placeholder={'******************'} aria-required="true">
+            <BaseInputSlot onClick={() => {
+                setIsShowPassword((prev) => !prev)
+            }} aria-label={isShowPassword ? "Hide password" : "Show password"}>
+                {isShowPassword ? <EyeOutline/> : <EyeOffOutline/>}
+            </BaseInputSlot>
+        </BaseInput>}
+
+            {type === 'search' && <BaseInput type={'text'} {...rest}>
+                <BaseInputSlot>
+                    <SearchOutline stroke='currentColor'/>
+                </BaseInputSlot>
+            </BaseInput>}
+    </>
+}
